requirements.txt
# Core
python-dotenv>=1.0.0

# Git ingestion
gitpython>=3.1.41

# LangGraph orchestration
langgraph>=0.2.0
langchain-core>=0.2.0
langchain-anthropic>=0.1.0

# PydanticAI synthesis
pydantic-ai>=0.0.9
pydantic>=2.7.0

# LLM
openai>=1.0.0
anthropic>=0.28.0

# Dashboard
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0

# Utilities
rich>=13.7.0
tqdm>=4.66.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
